removed broken functionalites and cleaned new one
